First commit with possibly very buggy script.
Script has a few escapes that mightn't be needed. For example,
the "apt-get install" line is broken over several lines so each
package has its own line, for clarity. "gcc" is escaped because
it gets syntax-highlighted, but this mightn't be necessary.

I performed more or less the exact steps in this script when
I installed PyGame, and it worked. If you get errors, and
especially if you *resolve* those errors, please share your
results.